ces-ddr-platform/ces-backend/pyproject.toml
[project]
name = "ces-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4,<2.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "bcrypt>=4.0.0,<5.0.0",
    "fastapi>=0.135.1,<0.136.0",
    "httpx>=0.28.1,<0.29.0",
    "psycopg[binary]>=3.2.0,<4.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "python-decouple>=3.8,<4.0",
    "python-jose[cryptography]>=3.5.0,<4.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "uvicorn[standard]>=0.38.0,<0.39.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3,<10.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
